Finishitem

+++ db/migrate/20230522071145_create_items.rb
@@ -4,7 +4,7 @@ class CreateItems < ActiveRecord::Migration[6.1]
       t.integer :genre_id
       t.string :name
       t.text :explanation
-      t.decimal :price_without_tax
+      t.integer :price_without_tax
       t.boolean :sales_status
       t.timestamps
     end
